including layer on cognito trigger, increase memory (memory was the issue)
