+++ web-api/terraform/api/outputs.tf
@@ -11,3 +11,7 @@ output "api_custom" {
 output "websockets_domain" {
   value = aws_apigatewayv2_domain_name.websockets_domain
 }
+
+output "puppeteer_layer_arn" {
+  value = aws_lambda_layer_version.puppeteer_layer.arn
+}
